Add the Dashboard hints to the image file.
If I understand Paul correctly, if one is going to restore the
Dashboard, one needs the cold registrations, as one really just
needs the assocation between a cell and the registrations which
were associated with it.

The real trick is in thawing this data when reading the image back
into memory.