Load back at least the gate.
This loads a nock value out of the formula and at least in theory,
loads the cold registrations back out. We can run the gate loaded
from an image, but I don't appear to be restoring the cold jet
associations correctly, as we never run the mock jetted decrement
function.