Snapshoting works.
This is able to boot a fakefed, save an image, and then restore
it on restart.

Caveats: saving an image is a bit slow and will probably get
slower over time. The java process hangs around for awhile
on shutdown performing the serialization and we'll want
to make the daemon wait for it to send an all clear.