Basic argon2 implementation taken from BouncyCastle.
BouncyCastle contains an implementation of Argon2, but Urbit's
version of argon2 is customized with a non-standard %u mode. This
takes the upstream code and adds the custom %u mode.

This also adds a test of the %u mode. This does not hook things up
to the jet dashboard, which is next.